Update Space home layout
